feat(#5): Props로 날씨 데이터 전달
- Home 컴포넌트에서 날씨 데이터 관리하며 props로 필요한 컴포넌트에 전달
- type 파일에서 필요한 타입 모듈 임포트
