+++ src/pages/home.tsx
@@ -2,13 +2,24 @@ import styled from "styled-components";
 import CityName from "@/components/city-name";
 import Temperature from "@/components/temperature";
 import TodayDate from "@/components/date";
+import { useCurrentLocation } from "@/hooks/useGeoLocation";
+import useWeather from "@/hooks/useWeather";
 
-const Home = () => {
+const geolocationOptions = {
+  enableHighAccuracy: true,
+  timeout: 1000 * 10, // 10초
+  maximumAge: 1000 * 3600 * 24, // 24시간
+};
+
+const Home: React.FC = () => {
+  const { loc } = useCurrentLocation(geolocationOptions);
+  const { data, isLoading, error } = useWeather(loc?.latitude, loc?.longitude);
+
   return (
     <Container className="mx-auto my-auto">
       <TodayDate />
-      <CityName />
-      <Temperature />
+      <CityName weatherData={data} isLoading={isLoading} error={error} />
+      <Temperature weatherData={data} isLoading={isLoading} error={error} />
     </Container>
   );
 };
